+++ packages/core/src/public_api.ts
@@ -7,3 +7,4 @@ export * from './validator-loader.service';
 export * from './validator.class';
 export * from './directives/form-validator.directive';
 export * from './directives/form-submit.directive';
+export * from './strategies';