+++ packages/core/src/strategies/noop-validation-display-strategy.ts
@@ -1,0 +1,8 @@
+import { IValidationDisplayStrategy } from './validation-display-strategy';
+
+export class NoopValidationDisplayStrategy implements IValidationDisplayStrategy {
+    /** Does nothing, as this validation message display strategy is a no-op. */
+    showError(element: HTMLElement, errorMessages: string[]): void {}
+
+    removeError(element: HTMLElement): void {}
+}