+++ packages/core/src/validator.class.ts
@@ -1,22 +1,18 @@
 import { InjectionToken } from '@angular/core';
+import { IValidationDisplayStrategy } from './strategies';
 
 export interface Dictionary<T> {
     [key: string]: T;
 }
 
-export declare type NgxFormValidationMessages = Dictionary<Dictionary<string>>;
+export declare type NgxValidationMessages = Dictionary<Dictionary<string>>;
 
-export interface NgxFormValidatorConfig {
-    showElementError?:
-        | boolean
-        | ((element: HTMLElement, errorMessages: string[]) => void);
-    removeElementError?: boolean | ((element: HTMLElement) => void);
-    validationMessages?: NgxFormValidationMessages;
+export interface NgxValidatorConfig {
+    validationDisplayStrategy?: IValidationDisplayStrategy;
+    validationMessages?: NgxValidationMessages;
 }
-export interface NgxFormValidatorGlobalConfig extends NgxFormValidatorConfig {
+export interface NgxValidatorGlobalConfig extends NgxValidatorConfig {
     globalValidationMessages?: Dictionary<string>;
 }
 
-export const NGX_VALIDATOR_CONFIG = new InjectionToken<
-NgxFormValidatorGlobalConfig
->('NGX_VALIDATION_CONFIG');
+export const NGX_VALIDATOR_CONFIG = new InjectionToken<NgxValidatorGlobalConfig>('NGX_VALIDATION_CONFIG');