+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3
+
+Add validation display strategy.
+
+Break changes:
+1. rename `NgxFormValidatorConfig` -> `NgxValidatorConfig`;
+1. rename `NgxFormValidatorGlobalConfig` -> `NgxValidatorGlobalConfig`;
+1. remove `NgxValidatorConfig` config 's `showError` and `removeError` options, add `validationDisplayStrategy` option replace it, support bootstrap strategy and noop strategy, default is bootstrap, you can add your own strategy through implement `IValidationDisplayStrategy`.
+
 ## 0.0.2
 1. 初始化 `NgxValidatorModule` 模块；
 1. 添加 `ngx-form-validator`（`ngxFormValidator`） 指令，主要在 ngForm 上追加一些验证相关的行为；